+++ src/provider/client/types/was/serviceResponse.ts
@@ -1,6 +1,0 @@
-export type ServiceResponse<DataType> = {
-  responseCode?: string;
-  count?: number;
-  hasMoreRecords?: boolean;
-  data?: DataType;
-};