+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Responses with large bodies would hang the program
+
+### Added
+
+- Additional Qualys API response error handling
+
 ## 4.7.1 - 2020-12-03
 
 ### Fixed